requirements/requirements.txt
apricot-select>=0.6.0
numba>=0.43.0
scipy>=1.5.0
scikit-learn
torch>=1.8.0
torchvision
torchtext
redis
tqdm>=4.24.0
sphinxcontrib-napoleon
sphinxcontrib-bibtex
sphinx-rtd-theme
matplotlib
numpy>=1.19.0
pillow>=8.4.0
pandas>=1.1.0
scikit-image>=0.17.0
pyyaml>=5.4.1
dotmap>=1.3.24
setuptools>=58.0.4
ray[tune]
ray[default]
datasets
pyJoules
wandb
pynvml
torchmetrics
seaborn
gdown
umap-learn